Use boost hash throughout
